requirements.txt
modal
pydantic==2.10.4
python-dotenv
duckdb
requests